# src/config/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Insomnia",
		"description": "",
		"contact": {},
		"version": "1.0"
	},
	"servers": [
		{
			"url": "/api",
			"variables": {}
		}
	],
	"paths": {
...
		"/v1/posts": {
			"get": {
				"tags": ["Post"],
				"summary": "Get All Posts",
				"description": "",
				"operationId": "GetAllPosts",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"text/plain": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				},
				"deprecated": false
			},
			"post": {
				"tags": ["Post"],
				"summary": "Create Post",
				"description": "",
				"operationId": "CreatePost",
				"parameters": [],
				"requestBody": {
					"description": "",
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostRequest"
							},
							"example": {
								"title": "primer post",
								"content": "dsfuhdasuhf udshf adh huidhsaf usdh fhusd",
								"author": "5e9806edbf2acd22b0cb0eb3"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "",
						"headers": {},
						"content": {
							"text/plain": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				},
				"deprecated": false
			}
		},
...
	},
	"components": {
		"schemas": {
			"CreatePostRequest": {
				"title": "CreatePostRequest",
				"required": ["title", "content", "author"],
				"type": "object",
				"properties": {
					"title": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"author": {
						"type": "string"
					}
				},
				"example": {
					"title": "primer post",
					"content": "dsfuhdasuhf udshf adh huidhsaf usdh fhusd",
					"author": "5e9806edbf2acd22b0cb0eb3"
				}
			},
...
		},
		"securitySchemes": {
			"httpBearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"security": [
		{
			"httpBearer": []
		}
	],
...
}